requests involving updating users now checks if the user exists first

+++ Server/src/Routes/DeleteUser.java
@@ -32,21 +32,26 @@ public class DeleteUser {
         } else {
             // check if token from parameters is valid and if session token has not yet expired
             if (sessions.containsKey(parameters.get(1)) && !hasTokenExpired(sessions, parameters.get(1))) {
-                // check if the user has the "Edit Users" permission
-                if (Users.userHasPermission(sessions.get(parameters.get(1)).get(0), ServerPermissions.EDIT_USERS)) {
-                    // check if the user is trying to delete themselves
-                    if (parameters.get(0).equals(sessions.get(parameters.get(1)).get(0))) {
-                        oos.writeObject(new Response(StatusCodes.FORBIDDEN, "Users Cannot Delete Themselves"));
-                    } else {
-                        // delete user
-                        if (Users.deleteUserFromDB(parameters.get(0))) {
-                            oos.writeObject(new Response(StatusCodes.OK, "User Deletion Successful"));
+                // check if the user exists
+                if (Users.doesUserExists(parameters.get(0))) {
+                    // check if the user has the "Edit Users" permission
+                    if (Users.userHasPermission(sessions.get(parameters.get(1)).get(0), ServerPermissions.EDIT_USERS)) {
+                        // check if the user is trying to delete themselves
+                        if (parameters.get(0).equals(sessions.get(parameters.get(1)).get(0))) {
+                            oos.writeObject(new Response(StatusCodes.FORBIDDEN, "Users Cannot Delete Themselves"));
                         } else {
-                            oos.writeObject(new Response(StatusCodes.INTERNAL_ERROR, "User Deletion Unsuccessful"));
+                            // delete user
+                            if (Users.deleteUserFromDB(parameters.get(0))) {
+                                oos.writeObject(new Response(StatusCodes.OK, "User Deletion Successful"));
+                            } else {
+                                oos.writeObject(new Response(StatusCodes.INTERNAL_ERROR, "User Deletion Unsuccessful"));
+                            }
                         }
+                    } else {
+                        oos.writeObject(new Response(StatusCodes.FORBIDDEN, "Missing Permissions"));
                     }
                 } else {
-                    oos.writeObject(new Response(StatusCodes.FORBIDDEN, "Missing Permissions"));
+                    oos.writeObject(new Response(StatusCodes.INTERNAL_ERROR, "User Does Not Exist"));
                 }
             } else {
                 oos.writeObject(new Response(StatusCodes.UNAUTHORISED, "Unauthorised Request"));
